feat: List CollectionView UI 최종구현

+++ OpenMarket/OpenMarket/View/TextAttribute/String+Extension.swift
@@ -1,4 +1,3 @@
-import Foundation
 import UIKit
 
 extension String {
@@ -32,7 +31,7 @@ extension String {
 
 extension String {
     var decimal: String? {
-        guard let number = Int(self) else {
+        guard let number = Double(self) else {
             return nil
         }
         let numberformatter = NumberFormatter()
